+++ R/download_s2_geom_acq.R
@@ -3,7 +3,7 @@
 #' @param aoi simple feature. vector including all S2 tiles intersecting with aoi
 #' @param geom_dir character. directory where rasters for geometry are saved
 #' @param nbCPU numeric. number of CPUs
-#' @param collection_GeomAcq_S2 list. STAC collection
+#' @param collection_geom_acq_S2 list. STAC collection
 #'
 #' @return list of collections per plot
 #' @importFrom CDSE GetOAuthToken GetImage
@@ -15,7 +15,7 @@
 #' @importFrom progressr with_progress progressor handlers
 #' @export
 
-download_s2_geom_acq <- function(aoi, geom_dir, collection_GeomAcq_S2, nbCPU=1){
+download_s2_geom_acq <- function(aoi, geom_dir, collection_geom_acq_S2, nbCPU=1){
 
   collection <- 'sentinel-2-l2a'
   # get token for authentication on CDSE
@@ -38,7 +38,7 @@ download_s2_geom_acq <- function(aoi, geom_dir, collection_GeomAcq_S2, nbCPU=1){
     # download geometry of acquisition for full scene (1km resolution)
     script_geom <- system.file("extdata", "S2L2A_geometry.js", package = "preprocS2")
 
-    DatesofAcq <- unique(get_dateAcq(S2product = collection_GeomAcq_S2$sourceId,
+    DatesofAcq <- unique(get_dateAcq(S2product = collection_geom_acq_S2$sourceId,
                                      stac_info = list('provider' = 'esa')))
     fileName <- fileNameAll <- list()
     if (nbCPU==1){